abstracted selectors into helpers component
